wip: remove pokemon-data.json
