+++ app/models/colyseus-models/pokemon.ts
@@ -47,7 +47,6 @@ import { distanceM } from "../../utils/distance"
 import { coinflip, pickRandomIn } from "../../utils/random"
 import { values } from "../../utils/schemas"
 import PokemonFactory from "../pokemon-factory"
-import { getPokemonData } from "../precomputed"
 import Player from "./player"
 
 export class Pokemon extends Schema implements IPokemon {
@@ -81,7 +80,7 @@ export class Pokemon extends Schema implements IPokemon {
   canHoldItems = true
   stages?: number
 
-  constructor(shiny: boolean, emotion: Emotion) {
+  constructor(shiny = false, emotion = Emotion.NORMAL) {
     super()
     const name = Object.entries(PokemonClasses).find(
       ([name, pokemonClass]) => pokemonClass === this.constructor
@@ -149,7 +148,7 @@ export class Pokemon extends Schema implements IPokemon {
 
   isInRegion(pkm: Pkm, map: DungeonPMDO, state?: GameState) {
     const regionSynergies = DungeonDetails[map]?.synergies
-    return regionSynergies.some((s) => getPokemonData(pkm).types.includes(s))
+    return regionSynergies.some((s) => new PokemonClasses[pkm]().types.has(s))
   }
 }
 
@@ -13401,8 +13400,8 @@ export class Tangrowth extends Pokemon {
 export const PokemonClasses: Record<
   Pkm,
   new (
-    shiny: boolean,
-    emotion: Emotion
+    shiny?: boolean,
+    emotion?: Emotion
   ) => Pokemon
 > = {
   [Pkm.DEFAULT]: Pokemon,
