+++ app/public/src/game/components/emote-menu.tsx
@@ -2,7 +2,7 @@ import { GameObjects } from "phaser"
 import React from "react"
 import ReactDOM from "react-dom/client"
 import { useTranslation } from "react-i18next"
-import { PRECOMPUTED_EMOTIONS_PER_POKEMON_INDEX } from "../../../../models/precomputed"
+import { PRECOMPUTED_EMOTIONS_PER_POKEMON_INDEX } from "../../../../models/precomputed/precomputed-emotions"
 import { IPlayer } from "../../../../types"
 import { AvatarEmotions, Emotion } from "../../../../types/enum/Emotion"
 import { logger } from "../../../../utils/logger"
