+++ app/public/src/game/components/pokemon-detail.tsx
@@ -2,7 +2,7 @@ import { t } from "i18next"
 import { GameObjects } from "phaser"
 import React from "react"
 import ReactDOM from "react-dom/client"
-import { getPokemonData } from "../../../../models/precomputed"
+import { getPokemonData } from "../../../../models/precomputed/precomputed-pokemon-data"
 import { Emotion } from "../../../../types"
 import { RarityColor } from "../../../../types/Config"
 import { Ability } from "../../../../types/enum/Ability"
