+++ app/public/src/pages/component/collection/pokemon-emotions-modal.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useMemo } from "react"
 import Modal from "react-bootstrap/esm/Modal"
 import { useTranslation } from "react-i18next"
-import { PRECOMPUTED_EMOTIONS_PER_POKEMON_INDEX } from "../../../../../models/precomputed"
+import { PRECOMPUTED_EMOTIONS_PER_POKEMON_INDEX } from "../../../../../models/precomputed/precomputed-emotions"
 import { Emotion } from "../../../../../types"
 import { Pkm, PkmIndex } from "../../../../../types/enum/Pokemon"
 import { useAppDispatch, useAppSelector } from "../../../hooks"
