+++ app/public/src/pages/component/game/game-pokemon-detail.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from "react"
 import { useTranslation } from "react-i18next"
 import { Pokemon } from "../../../../../models/colyseus-models/pokemon"
 import PokemonFactory from "../../../../../models/pokemon-factory"
-import { getPokemonData } from "../../../../../models/precomputed"
+import { getPokemonData } from "../../../../../models/precomputed/precomputed-pokemon-data"
 import { Emotion } from "../../../../../types"
 import { RarityColor } from "../../../../../types/Config"
 import { Ability } from "../../../../../types/enum/Ability"
