+++ app/public/src/pages/component/wiki/wiki-pokemons.tsx
@@ -5,7 +5,7 @@ import { RarityColor } from "../../../../../types/Config"
 import { Rarity } from "../../../../../types/enum/Game"
 import { PokemonTypeahead } from "../typeahead/pokemon-typeahead"
 import { Pkm, PkmFamily, PkmIndex } from "../../../../../types/enum/Pokemon"
-import { getPokemonData } from "../../../../../models/precomputed"
+import { getPokemonData } from "../../../../../models/precomputed/precomputed-pokemon-data"
 import { PRECOMPUTED_POKEMONS_PER_RARITY } from "../../../../../models/precomputed/precomputed-rarity"
 import { getPortraitSrc } from "../../../utils"
 import WikiPokemonDetail from "./wiki-pokemon-detail"
