+++ package.json
@@ -52,7 +52,7 @@
     "@reduxjs/toolkit": "^2.2.1",
     "@types/bootstrap": "^5.2.10",
     "body-parser": "^1.20.1",
-    "colyseus": "^0.15.15",
+    "colyseus": "^0.15.17",
     "colyseus.js": "^0.15.16",
     "compression": "^1.7.4",
     "cors": "^2.8.5",
